[Feat] Animation Features
- Animations can be played on loop or once only
- Animations can be set to be played in reverse

+++ Engine/core-engine/src/Editor/editor-hierarchy-list.h
@@ -86,7 +86,7 @@ namespace Copium
 			if a game object is selected return true
 		*/
 		/*******************************************************************************/
-		bool display_gameobject(GameObject& _go, GameObjectID& _selected, std::vector<GameObject*>& _vector, int _index);
+		bool display_gameobject(GameObject& _go, GameObjectID& _selected, std::vector<GameObject*>& _vector, size_t _index);
 		/*******************************************************************************
 		/*!
 		*
@@ -115,7 +115,7 @@ namespace Copium
 			if a game object is selected return true
 		*/
 		/*******************************************************************************/
-		bool display_gameobject(GameObject& _go, GameObjectID& _selected, std::list<Transform*>& _list, int _index);
+		bool display_gameobject(GameObject& _go, GameObjectID& _selected, std::list<Transform*>& _list, size_t _index);
 		/*******************************************************************************
 		/*!
 		*
